pyproject.toml
[tool.poetry]
name = "heekkr-resolver-simple"
version = "0.1.0"
description = ""
authors = ["jangjunha <hi@jangjunha.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.2"
aiohttp = "^3.8.5"
aiodns = "^3.0.0"
heekkr = {version = "1.2.0", source = "gar"}
lxml = "^4.9.3"
aiocache = "^0.12.2"
aiostream = "^0.4.5"
openpyxl = "^3.1.2"
google-cloud-storage = "^2.10.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.284"
ruff-lsp = "^0.0.35"
pyright = "^1.1.322"
types-cachetools = "^5.3.0.6"
black = "^23.7.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"


[[tool.poetry.source]]
name = "gar"
url = "https://asia-northeast3-python.pkg.dev/heekkr/heekkr-python/simple/"
priority = "explicit"

[tool.ruff]
target-version = "py311"
[tool.ruff.extend-per-file-ignores]
"tests/**/*.py" = ["E501"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
